+++ identity_from_directory.php
@@ -35,16 +35,17 @@ class identity_from_directory extends rcube_plugin
     {
         $this->load_config('config.inc.php.dist'); // load the plugin's distribution config file as default
         $this->load_config(); // merge with local configuration file (which can overwrite any settings)
-
         if ($this->ldap) {
             return $args;
         }
+        $debug_plugin = (bool) $this->rc->config->get('identity_from_directory_debug');
 
         // get user's data from directory and prepare it for further processing
-        $user_data = array();
-        $debug_plugin = (bool) $this->rc->config->get('identity_from_directory_debug');
-        $ad_handle_proxyaddresses = (bool) $this->rc->config->get('identity_from_directory_handle_proxyaddresses');
-        $exclude_alias_regex = (string) $this->rc->config->get('identity_from_directory_exclude_alias_regex');
+        $user_data = [
+            'name' => '',
+            'email' => '', // default / main email address; IDN domains (no Punycode/ACE)
+            'email_list' => [], // list of all email addresses (main, default, aliases); IDN domains (no Punycode/ACE)
+        ];
         if ($this->init_ldap([
             'mail_host' => $this->rc->user->data['mail_host'],
         ])) {
@@ -56,64 +57,67 @@ class identity_from_directory extends rcube_plugin
             $results = $this->ldap->search('*', $this->rc->user->data['username'], true);
 
             if (count($results->records) === 1) {
+
+                $ad_handle_proxyaddresses = (bool) $this->rc->config->get('identity_from_directory_handle_proxyaddresses');
+                $exclude_alias_regex = (string) $this->rc->config->get('identity_from_directory_exclude_alias_regex');
+
                 $ldap_entry = $results->records[0];
                 if ($debug_plugin) {
-                    rcube::write_log('identity_from_directory', 'Found a record for' . $this->rc->user->data['username'] . ': '.print_r($ldap_entry, true));
+                    rcube::write_log('identity_from_directory',
+                        'Found a record for ' . $this->rc->user->data['username'] . ': '
+                        . print_r($ldap_entry, true));
                 }
 
-                $user_data['user_name'] = is_array($ldap_entry['name']) ? $ldap_entry['name'][0] : $ldap_entry['name'];
-                $user_data['user_email'] = is_array($ldap_entry['email']) ? $ldap_entry['email'][0] : $ldap_entry['email'];
-                $user_data['email_list'] = [];
-                if (strpos($user_data['user_email'], '@') !== false) {
-                    $user_data['user_email'] = rcube_utils::idn_to_ascii($user_data['user_email']);
+                $user_data['name'] = is_array($ldap_entry['name']) ? $ldap_entry['name'][0] : $ldap_entry['name'];
+                $user_data['email'] = rcube_utils::idn_to_utf8(trim(is_array($ldap_entry['email']) ? $ldap_entry['email'][0] : $ldap_entry['email']));
+                if (!empty($user_data['email'])) {
+                    $user_data['email_list'][] = $user_data['email'];
                 }
-                if (!empty($user_data['user_email'])) {
-                    $user_data['email_list'][] = $user_data['user_email'];
-                }
-                $user_data['email_default'] = $user_data['user_email'];
 
                 foreach (array_keys($ldap_entry) as $key) {
                     // add email addresses (main, aliases) to the list for the user
-                    if (preg_match('/^email($|:)/', $key)) {
+                    if (preg_match('/^email($|:)/i', $key)) {
                         foreach ((array) $ldap_entry[$key] as $alias) {
-                            $alias = trim($alias);
-                            if (empty($alias) || in_array($alias, $user_data['email_list'])) {
+                            $alias = rcube_utils::idn_to_utf8(trim($alias));
+                            if (empty($alias) || self::email_in_array($alias, $user_data['email_list'])) {
                                 continue;
                             }
                             if (strpos($alias, '@') === false || (!empty($exclude_alias_regex) && (preg_match($exclude_alias_regex, $alias)))) {
                                 if ($debug_plugin) {
-                                    rcube::write_log('identity_from_directory', 'Excluded '. $alias . ' from handling as it is an invalid email address or matching "'. $exclude_alias_regex .'" (identity_from_directory_exclude_alias_regex).');
+                                    rcube::write_log('identity_from_directory',
+                                        'Excluded ' . $alias . ' from handling as it is an invalid email address or matching "'
+                                        . $exclude_alias_regex . '" (identity_from_directory_exclude_alias_regex).');
                                 }
                                 continue;
                             }
-                            $user_data['email_list'][] = rcube_utils::idn_to_ascii($alias);
+                            $user_data['email_list'][] = $alias;
                         }
-                    } elseif ($ad_handle_proxyaddresses &&
-                              preg_match('/^proxyaddresses($|:)/', $key)) {
-                        // Handle Active Directory attribute proxyAddresses, originally a CSV string
-                        // like 'smtp:foo@exmaple.com,bar@example.net'. Returned as string if there
-                        // is only one, returned as array if there are multiple.
+
+                    // handle Active Directory attribute "proxyAddresses"
+                    } elseif ($ad_handle_proxyaddresses && preg_match('/^proxyaddresses($|:)/i', $key)) {
+                        // originally a CSV string like 'smtp:foo@exmaple.com,bar@example.net'.
+                        // The used library returns as string if there is only one address and
+                        // an array if there are multiple.
                         $proxyaddresses = $ldap_entry[$key];
                         if (!is_array($proxyaddresses)) {
                             $proxyaddresses = [ $proxyaddresses ];
                         }
                         foreach ((array) $proxyaddresses as $alias) {
-                            $alias = trim($alias);
-                            if (empty($alias)) {
-                                continue;
-                            }
-                            $alias = trim(preg_replace('/^smtp:(.+)/', '\1', $alias, 1));
-                            if (empty($alias) || in_array($alias, $user_data['email_list'])) {
+                            $alias = rcube_utils::idn_to_utf8(trim(preg_replace('/^smtp:(.+)/i', '\1', trim($alias), 1)));
+                            if (empty($alias) || self::email_in_array($alias, $user_data['email_list'])) {
                                 continue;
                             }
                             if (strpos($alias, '@') === false || (!empty($exclude_alias_regex) && (preg_match($exclude_alias_regex, $alias)))) {
                                 if ($debug_plugin) {
-                                    rcube::write_log('identity_from_directory', 'Excluded '. $alias . ' from handling as it is an invalid email address or matching "'. $exclude_alias_regex .'" (identity_from_directory_exclude_alias_regex).');
+                                    rcube::write_log('identity_from_directory',
+                                        'Excluded '. $alias . ' from handling as it is an invalid email address or matching "'
+                                        . $exclude_alias_regex . '" (identity_from_directory_exclude_alias_regex).');
                                 }
                                 continue;
                             }
-                            $user_data['email_list'][] = rcube_utils::idn_to_ascii($alias);
+                            $user_data['email_list'][] = $alias;
                         }
+
                     // add LDAP data but exclude _ID, _raw_attrib etc. and do not overwrite already existing keys
                     } elseif (strpos($key, '_') !== 0 && !array_key_exists($key, $user_data)) {
                         $user_data[$key] = $ldap_entry[$key];
@@ -122,17 +126,17 @@ class identity_from_directory extends rcube_plugin
                 $user_data['email_list'] = array_unique($user_data['email_list']);
 
             } elseif ($debug_plugin && count($results->records) > 1) {
-               rcube::write_log('identity_from_directory', 'Searching for ' . $this->rc->user->data['username'] . ' returned more than one result, all were ignored as unambiguous assignment is not possible.');
+                rcube::write_log('identity_from_directory',
+                    'Searching for '. $this->rc->user->data['username']
+                    . ' returned more than one result; all were ignored as unambiguous assignment is not possible.');
             }
         }
-        if (!array_key_exists('email_list', $user_data) || empty($user_data['email_list'])) {
+        if (empty($user_data['email_list'])) {
             return $args;
         }
 
         // get config and other data needed for further processing
         $ldap_config = (array) $this->rc->config->get('identity_from_directory_ldap');
-        $delete_unmanaged = (bool) $this->rc->config->get('identity_from_directory_delete_unmanaged');
-        $exclude_delete_unmanaged_regex = (string) $this->rc->config->get('identity_from_directory_exclude_delete_unmanaged_regex');
         $update_signatures = (bool) $this->rc->config->get('identity_from_directory_update_signatures');
         $use_html_signature = (bool) $this->rc->config->get('identity_from_directory_use_html_signature');
         $wash_html_signature = (bool) $this->rc->config->get('identity_from_directory_wash_html_signature');
@@ -142,7 +146,7 @@ class identity_from_directory extends rcube_plugin
             $signature_template = (string) $this->rc->config->get('identity_from_directory_signature_template_plaintext');
         }
         $signature_fallback_values = (array) $this->rc->config->get('identity_from_directory_fallback_values');
-        $identities = $this->rc->user->list_emails();
+        $identities_existing = $this->rc->user->list_emails(); // list of all user emails (from identities), array with identity_id, name and email address
 
         // maintain an identity for each of the user's determined email addresses
         foreach ((array) $user_data['email_list'] as $email) {
@@ -151,15 +155,17 @@ class identity_from_directory extends rcube_plugin
             $is_standard = 0; // 1: use the identity as default (there can only be one)
             $signature   = $signature_template; // copy signature template
 
-            foreach ($identities as $identity) {
-                if ($identity['email'] === $email) {
+            foreach ($identities_existing as $identity_existing) {
+                // case-insensitive search to update an existing identity, even if
+                // there are differences in capitalization.
+                if (self::email_in_array($identity_existing['email'], [ $email ])) {
                     $hook_to_use = 'identity_update';
-                    $identity_id = $identity['identity_id'];
+                    $identity_id = $identity_existing['identity_id'];
                     break;
                 }
             }
 
-            if (strtolower($user_data['email_default']) === strtolower($email)) {
+            if ($user_data['email'] === $email) {
                 $is_standard = 1;
             }
 
@@ -167,7 +173,7 @@ class identity_from_directory extends rcube_plugin
             $identity_record = [
                 'user_id' => $this->rc->user->ID,
                 'standard' => $is_standard,
-                'name' => (!empty($user_data['name']) ? $user_data['name'] : $user_data['user_name']),
+                'name' => (!empty($user_data['name']) ? $user_data['name'] : $this->rc->user->data['username']),
                 'email' => $email,
                 'organization' => (array_key_exists('organization', $user_data) ? $user_data['organization'] : ''),
             ];
@@ -177,14 +183,16 @@ class identity_from_directory extends rcube_plugin
                 // from LDAP or $config['identity_from_directory_fallback_values']:
                 // - %foo%: raw value of field 'foo'
                 // - %foo_html%: HTML entities encoded value of field 'foo'
-                // - %foo_url%: URL encoded value of field 'foo'
+                // - %foo_url%: URL encoded value of field 'foo'. Additional optimizations are
+                //   applied for the fields 'email' (usage of Punycode for email domains),
+                //   'phone' and 'fax' (stripping of chars not compatible with tel:// URLs)
                 foreach (array_keys($ldap_config['fieldmap']) as $fieldmap_key) {
                     $replace_raw = '';
                     if ($fieldmap_key === 'email') {
                         // Use the correct email address (alias) of the corresponding identity for
-                        // the %email%, %email_html% and %email_url% placeholders instead of the single
-                        // mapped value returned by the directory (which should be stored in
-                        // $user_data['user_email']). Otherwise, the same single email address value would
+                        // the %email%, %email_html% and %email_url% placeholders instead of the
+                        // single mapped value returned by the directory (which should be stored in
+                        // $user_data['email']). Otherwise, the same single email address value would
                         // be used for all of the user's identities (even the one of alias addresses).
                         $replace_raw = (string) $email;
                     } elseif (array_key_exists($fieldmap_key, $user_data) && ((string) $user_data[$fieldmap_key] !== '')) {
@@ -198,21 +206,25 @@ class identity_from_directory extends rcube_plugin
                     }
 
                     $replace_html = '';
-                    $replace_html = htmlspecialchars($replace_raw, ENT_NOQUOTES, RCUBE_CHARSET);
+                    $replace_html = htmlspecialchars($replace_raw, \ENT_NOQUOTES, RCUBE_CHARSET);
 
                     $replace_url = '';
                     if ($fieldmap_key === 'phone' || $fieldmap_key === 'fax') {
-                        $replace_url = urlencode(preg_replace('/[^+0-9]+/', '', $replace_raw)); // strip some chars for "tel://" URL usage
+                        // strip some chars for "tel://" URL usage
+                        $replace_url = urlencode(preg_replace('/[^+0-9]+/', '', $replace_raw));
+                    } elseif ($fieldmap_key === 'email') {
+                        // use Punycode/ACE for "mailto://" URL usage
+                        $replace_url = urlencode(rcube_utils::idn_to_ascii($replace_raw));
                     } else {
                         $replace_url = urlencode($replace_raw);
                     }
 
                     $signature = str_replace([ '%'. $fieldmap_key . '%',
                                                '%'. $fieldmap_key . '_html%',
-                                               '%'. $fieldmap_key . '_url%', ],
-                                             [  $replace_raw,
-                                                $replace_html,
-                                                $replace_url, ], $signature);
+                                               '%'. $fieldmap_key . '_url%' ],
+                                             [ $replace_raw,
+                                               $replace_html,
+                                               $replace_url ], $signature);
                 }
 
                 $identity_record['html_signature'] = ($use_html_signature) ? 1 : 0;
@@ -234,31 +246,73 @@ class identity_from_directory extends rcube_plugin
         }
 
         // delete identities which are not managed by this plugin
+        $delete_unmanaged = (bool) $this->rc->config->get('identity_from_directory_delete_unmanaged');
+        $exclude_delete_unmanaged_regex = (string) $this->rc->config->get('identity_from_directory_exclude_delete_unmanaged_regex');
         if ($delete_unmanaged) {
-            $identities_count = count($identities);
-            foreach ($identities as $identity) {
-                if ($identities_count > 1 && !(in_array($identity['email'], $user_data['email_list']))) {
+            $identity_existing_count = count($identities_existing);
+            foreach ($identities_existing as $identity_existing) {
 
-                    if (!empty($exclude_delete_unmanaged_regex) && preg_match($exclude_delete_unmanaged_regex, $identity['email'])) {
+                if ($identity_existing_count > 1 &&
+                    // do NOT normalize email address here (e.g. via mb_strtolower(), trim(), and/or
+                    // rcube_utils::idn_to_utf8()). Otherwise, some identities to be cleaned up
+                    // would not be detected (e.g. leftovers with differences in only upper or lower
+                    // case characters or Punycode/ACE in domain names)
+                    !(in_array($identity_existing['email'], $user_data['email_list']))) {
+                    if (!empty($exclude_delete_unmanaged_regex) && preg_match($exclude_delete_unmanaged_regex, $identity_existing['email'])) {
                         if ($debug_plugin) {
-                            rcube::write_log('identity_from_directory', 'Excluded identity '. $identity['identity_id'] .' of user '.  $this->rc->user->data['username'] .' from automatic deletion. It\'s email '. $identity['email'] .' is not listed in the directory but matching "'. $exclude_delete_unmanaged_regex .'" (identity_from_directory_exclude_delete_unmanaged_regex).');
+                            rcube::write_log('identity_from_directory',
+                                'Excluded identity ' . $identity_existing['identity_id'] . ' of user '
+                                . $this->rc->user->data['username'] . ' from automatic deletion. It\'s email '
+                                . $identity_existing['email'] . ' is not listed in the directory but matching "'
+                                . $exclude_delete_unmanaged_regex
+                                . '" (identity_from_directory_exclude_delete_unmanaged_regex).');
                         }
                         continue;
                     }
 
                     if ($debug_plugin) {
-                        rcube::write_log('identity_from_directory', 'Deleting identity '. $identity['identity_id'] .' of user '.  $this->rc->user->data['username'] .' because it\'s email '. $identity['email'] .' is the not listed in the directory.');
+                        rcube::write_log('identity_from_directory',
+                            'Deleting identity '. $identity_existing['identity_id'] .' of user '
+                            . $this->rc->user->data['username'] .' because it\'s email '
+                            . $identity_existing['email'] . ' is the not listed in the directory.');
                     }
 
-                    if (!($this->rc->user->delete_identity($identity['identity_id'])) && $debug_plugin) {
-                        rcube::write_log('identity_from_directory', 'Could note delete identity '. $identity['identity_id'] .' for email '. $identity['email']);
+                    if (!($this->rc->user->delete_identity($identity_existing['identity_id'])) && $debug_plugin) {
+                        rcube::write_log('identity_from_directory',
+                            'Could note delete identity '. $identity_existing['identity_id']
+                            . ' for email '.$identity_existing['email']);
                     }
-                    $identities_count--;
+                    $identity_existing_count--;
                 }
+
             }
         }
 
         return $args;
+    }
+
+
+    /**
+     * Search for an email address in an array of email addresses. The search
+     * will ignores differences in capitalization or Punycode/ACE.
+     *
+     * RFC 5321 (Simple Mail Transfer Protocol) section 2.3.11 leaves it up
+     * to the host if the "local-part" in "local-part@domain" is case-
+     * insensitively. De-facto, it gets handled case-insensitive by most
+     * systems out there and users are expecting that Foo@example.com ==
+     * foo@example.com quite often. This function also acts like this.
+     *
+     * @param string $needle Value to seek.
+     * @param array $haystack Array to seek in.
+     * @return bool
+     */
+    public static function email_in_array($needle, $haystack)
+    {
+        $haystack_new = [];
+        foreach($haystack as $key => $value) {
+            $haystack_new[$key] = mb_strtolower(rcube_utils::idn_to_utf8(trim($value)), 'UTF-8');
+        }
+        return in_array(mb_strtolower(rcube_utils::idn_to_utf8(trim($needle)), 'UTF-8'), $haystack_new);
     }
 
 
@@ -292,7 +346,8 @@ class identity_from_directory extends rcube_plugin
             empty($ldap_config['search_fields']) ||
             empty($ldap_config['fieldmap'])) {
             if ($debug_plugin) {
-                rcube::write_log('identity_from_directory', 'The plugin config seems to be invalid, please check $config[\'identity_from_directory_ldap\'].');
+                rcube::write_log('identity_from_directory',
+                    'The plugin config seems to be invalid, please check $config[\'identity_from_directory_ldap\'].');
             }
             return false;
         }
@@ -300,7 +355,8 @@ class identity_from_directory extends rcube_plugin
             !array_key_exists('email', $ldap_config['fieldmap']) ||
             !array_key_exists('organization', $ldap_config['fieldmap'])) {
             if ($debug_plugin) {
-                rcube::write_log('identity_from_directory', 'The plugin config seems to be invalid, please check $config[\'identity_from_directory_ldap\'][\'fieldmap\'].');
+                rcube::write_log('identity_from_directory',
+                    'The plugin config seems to be invalid, please check $config[\'identity_from_directory_ldap\'][\'fieldmap\'].');
             }
             return false;
         }